refactor(e2e): write full integration
